onCommet, onCupid, onDamnFrickenDamage
I posted my issues with the onDamage() on bukkit forums and did get some help with it, but the issue with the "void is an invalid type for variable onDamage" error. I'll have more commits as the day progresses. Once the onDamage issue is done with, this will allow me to take out the potion effects on the other gems as well and make cleaner gem effects.